+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.2.0
+* Added support for linking multiple targets per SlotLink.
+* Due to the changed data-model, implemented an operator to migrate existing animations.\
+	It will have to be executed once, before working with SlotLink when SlotLink data already existed.\
+	No data gets lost, all will work as before.
+
 ## v0.1.6
 * The initial setup for a new SlotLink now tries to determine the target automatically.
 * Added button to setup the links for all Slots, and determine the target automatically if possible.